chore: move mod downloading to cli/cache
